#83 refactor: apiConfig 적용
